typesの追加や一覧の取得をDBから取得に修正
