+++ hooks/api/useCreateCombination.ts
@@ -1,15 +1,16 @@
-import { Combination } from "@/app/post/page";
+
+import { CombinationParms } from "@/types/combination";
 import { useAxios } from "../axios/useAxios";
 
 type CombinationRequest = {
-  combination: Combination;
+  combination: CombinationParms;
 };
 
 export const useCreateCombination = () => {
   const axios = useAxios();
 
   const createCombination = async (CombinationRequest: CombinationRequest) => {
-    const response = await axios.post("/combinations", CombinationRequest);
+    const response = await axios.postForm("/combinations", CombinationRequest);
     return response.data;
   };
