+++ types/combination.ts
@@ -1,0 +1,19 @@
+export type CombinationParms = {
+  title: string;
+  image: File | null;
+  flight: string;
+  shaft: string;
+  barrel: string;
+  tip: string;
+  description: string;
+};
+
+export type Combination = {
+  title: string;
+  image: string;
+  flight: string;
+  shaft: string;
+  barrel: string;
+  tip: string;
+  description: string;
+};
